@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	<TargetFramework>netcoreapp2.2</TargetFramework>
+	<TargetFramework>netcoreapp3.0</TargetFramework>
 	<RuntimeIdentifier>win7-x86</RuntimeIdentifier>
 	<AssemblyName>Havit.NewProjectTemplate.ApiModelsGenerator</AssemblyName>
 	<RootNamespace>Havit.NewProjectTemplate.ApiModelsGenerator</RootNamespace>
