@@ -3,6 +3,8 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Hosting;
+using Havit.NewProjectTemplate.WindsorInstallers;
 
 namespace Havit.NewProjectTemplate.WebAPI
 {
@@ -10,39 +12,41 @@ namespace Havit.NewProjectTemplate.WebAPI
     {
         public static void Main(string[] args)
         {
-            BuildWebHost(args).Run();
-        }
+			CreateHostBuilder(args).Build().Run();
+		}
 
-	    public static IWebHost BuildWebHost(string[] args)
-	    {
-		    return WebHost.CreateDefaultBuilder(args)
-			    .UseApplicationInsights()
-			    .UseStartup<Startup>()
+		public static IHostBuilder CreateHostBuilder(string[] args)
+		{
+			return Host.CreateDefaultBuilder(args)
+				.ConfigureWebHostDefaults(webBuilder =>
+				{
+					webBuilder
+						.UseStartup<Startup>()
 #if DEBUG
-			    .UseEnvironment("Development") // pro Red-Gate ANTS Performance Profiler
-			    .UseUrls("http://localhost:9901") // pro Red-Gate ANTS Performance Profiler
+						.UseEnvironment("Development") // pro Red-Gate ANTS Performance Profiler
+						.UseUrls("http://localhost:9901"); // pro Red-Gate ANTS Performance Profiler
 #endif
-			    .ConfigureAppConfiguration((hostContext, config) =>
-			    {
-				    // delete all default configuration providers
-				    config.Sources.Clear();
-				    config
-					    .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
-					    .AddJsonFile($"appsettings.{hostContext.HostingEnvironment.EnvironmentName}.json", optional: true, reloadOnChange: true)
-					    .AddJsonFile(@"Config\appsettings.json", optional: true, reloadOnChange: true)
-					    .AddEnvironmentVariables();
-			    })
-			    .ConfigureLogging((hostingContext, logging) =>
-			    {
-				    logging.AddConfiguration(hostingContext.Configuration.GetSection("Logging"));
-				    logging.AddConsole();
-				    logging.AddDebug();
+				})
+				.UseServiceProviderFactory(ctx => new WebApiServiceProviderFactory(ctx.Configuration))
+				.ConfigureAppConfiguration((hostContext, config) =>
+				{
+					// delete all default configuration providers
+					config.Sources.Clear();
+					config
+						.AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
+						.AddJsonFile($"appsettings.{hostContext.HostingEnvironment.EnvironmentName}.json", optional: true, reloadOnChange: true)
+						.AddJsonFile(@"Config\appsettings.json", optional: true, reloadOnChange: true)
+						.AddEnvironmentVariables();
+				})
+				.ConfigureLogging((hostingContext, logging) =>
+				{
+					logging.AddConfiguration(hostingContext.Configuration.GetSection("Logging"));
+					logging.AddConsole();
+					logging.AddDebug();
 #if !DEBUG
 					logging.AddEventLog();
 #endif
-			    })
-
-				.Build();
+				});
 	    }
     }
 }
