@@ -17,11 +17,11 @@ namespace Havit.NewProjectTemplate.WebAPI.Infrastructure.Tools
 		{
 			using (IServiceScope serviceScope = app.ApplicationServices.GetRequiredService<IServiceScopeFactory>().CreateScope())
 			{
-				var context = serviceScope.ServiceProvider.GetService<IDbContext>();
-				context.Database.Migrate();
+				//var context = serviceScope.ServiceProvider.GetService<IDbContext>();
+				//context.Database.Migrate();
 
-				var dataSeedRunner = serviceScope.ServiceProvider.GetService<IDataSeedRunner>();
-				dataSeedRunner.SeedData<CoreProfile>();
+				//var dataSeedRunner = serviceScope.ServiceProvider.GetService<IDataSeedRunner>();
+				//dataSeedRunner.SeedData<CoreProfile>();
 			}
 		}
 	}
